added default products data

+++ ecommerce-api/src/adapters/src/main/java/br/com/ecommerce/adapters/mapper/product/CreateProductInputData.java
@@ -10,7 +10,7 @@ public class CreateProductInputData implements InputMapper<Product> {
   private String description;
   private double price;
   private String image;
-  private Integer rate;
+  private Double rate;
   private Integer maxParcels;
   private Integer stock;
 
@@ -59,11 +59,11 @@ public class CreateProductInputData implements InputMapper<Product> {
     this.image = image;
   }
 
-  public Integer getRate() {
+  public Double getRate() {
     return rate;
   }
 
-  public void setRate(Integer rate) {
+  public void setRate(Double rate) {
     this.rate = rate;
   }
